Fix playlist ID deep link routing on mobile
