+++ packages/mobile/package.json
@@ -217,9 +217,7 @@
   },
   "jest": {
     "preset": "react-native",
-    "setupFilesAfterEnv": [
-      "@testing-library/jest-native/extend-expect"
-    ],
+    "setupFilesAfterEnv": ["<rootDir>/jest.setup.js"],
     "moduleFileExtensions": [
       "ts",
       "tsx",
